refactor: Components and utility functions
